requirements.txt
numpy
matplotlib
discord
requests
python-dotenv
pynacl
ffmpeg-python